cpp: Remove forward declarations for dependencies
Letting the compiler generate forward declarations for dependencies causes
problems when building shared libraries with MSVC where those forward
declarations inherit the current dllexport/dllimport symbol which might not be
the correct one as the dependency might be built in another shared library.

The generator already makes sure that the header files for all dependencies are
properly included in the current translation unit, header files which themselves
contain the forward declarations for this dependency, with proper export symbols
set.